requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.15.0
python-dateutil>=2.8.2
gspread>=6.1.0
google-auth>=2.29.0
google-api-python-client>=2.120.0
openpyxl>=3.1.0
